fix(agents): move the chain summary parser to the failover patterns module
CI caught user-copy.ts crossing its 700 line budget at 704. The leg parsing is
message pattern work, so it belongs with the other failover message patterns
rather than in the copy renderer, which puts the file back under budget at 692
without dropping anything from the fix.
